We can send data from contentscript.js to background.js now
